examples/custom-pipeline/pyproject.toml
[tool.poetry]
name = "simple-pipeline"
version = "0.0.1"
description = ""
authors = ["GenAI SDK Team <gat-sdk@gdplabs.id>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
gllm-pipeline-binary = "^0.2.17"
gllm-core-binary = "^0.2.15"
gllm-generation-binary = "^0.2.11"
gllm-rag-binary = "^0.0.2"
gllm-plugin-binary = "^0.0.6"

[project]
name = "simple-pipeline"
version = "0.0.1"
description = ""
authors = [
    { name = "GenAI SDK Team", email = "gat-sdk@gdplabs.id" }
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "gllm-core-binary>=0.2.15",
    "gllm-generation-binary>=0.2.11",
    "gllm-pipeline-binary>=0.2.17",
    "gllm-plugin-binary>=0.0.5",
    "gllm-rag-binary>=0.0.2",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "gllm-agents-binary (>=0.0.2,<0.0.3)",
]
